Fix for fragment download hang issue with some CDNs
We see that a CDN can sometimes stop returning bytes for a fragment.

The DASH-JS download progress logic is broken in this situation, as
it is solely driven from XHR progress events.  When the CDN is not
delivering data the ABR logic does not abandon the stalled fragment.

Additionally, DASH-JS does not have any value set in the xhr.timeout
property, which can act as a fallback should the abandonment rule(s)
fail.

A 20second xhr.timeout value seems like another prudent addition.